finetuning/requirements_finetune.txt
# Requirements for Qwen OCR Fine-tuning
torch>=2.1.0
transformers>=4.57.0
peft>=0.8.0
accelerate>=0.25.0
qwen-vl-utils>=0.0.14
pandas>=1.5.0
pillow>=9.0.0
datasets>=2.10.0
huggingface-hub>=0.20.0